+++ backend/src/services/agent-executor.ts
@@ -6,7 +6,7 @@ import { streamLLM, calcCost } from './llm-router'
 import { searchKnowledge } from './vector-search'
 import { parseDelegateDirectives, stripDelegateDirectives, executeDelegations, buildSynthesisPrompt } from './orchestrator'
 import { parseAgentWebhooks, stripAgentWebhooks, executeAgentWebhooks } from './outbound-webhooks'
-import { sendPushNotification } from '../routes/notifications'
+import { sendPushNotification } from './push'
 import { fireWebhook } from './outbound-webhooks'
 import { ensureFreshToken, searchDriveFiles } from './google-auth'
 import { isExternalAgent, notifyExternalAgent } from './agent-runtime'
